Add pin(note:) action
